Add solution for problem `C - Eugene and an array` Link: https://codeforces.com/contest/1333/problem/C Tags: binary search, data structures, implementation, two pointers, *1700 Ref: https://codeforces.com/contest/1333/submission/75920983

+++ codeforces/1333/C.cpp
@@ -24,7 +24,7 @@ typedef long long Long;
 typedef vector<int> vi;
 
 struct chash {
-  const int RANDOM =
+  const Long RANDOM =
       (long long)(make_unique<char>().get()) ^
       chrono::high_resolution_clock::now().time_since_epoch().count();
   static unsigned long long hash_f(unsigned long long x) {
